fix backup manager + add support for new official numbers

+++ src/data/manager/backup_manager.py
@@ -1,3 +1,4 @@
+import os.path
 import shutil
 import sys
 #from distutils.dir_util import copy_tree
@@ -15,21 +16,20 @@ def updateBackupConfig(sourceConfigPath = ''):
     else:
         shutil.copyfile('data/config.json', 'data/backup/config.json')
 
-def removeExistingBackupData():
-    # remove complete directory
-    shutil.rmtree('data/backup/data')
-
 def updateBackupDir(sourceConfigPath = ''):
     updateBackupConfig(sourceConfigPath)
-    removeExistingBackupData()
+    shutil.rmtree('data/backup/data') # delete data dir as well as its content
     shutil.copytree('data/data', 'data/backup/data')
 
 def repairData():
     trace.TRACE('REPAIRING DATA AND CONFIG')
     try:
         # repairing data + config
-        shutil.rmtree('data/data')
-        shutil.copytree('data/backup', 'data')
+        if (os.path.isdir('data/data')):
+            # delete data dir as well as its content
+            shutil.rmtree('data/data')
+        shutil.copytree('data/backup/data', 'data/data')
+        shutil.copyfile('data/backup/config.json', 'data/config.json')
     except Exception as e:
         trace.TRACE(e)
         # if repair failed -> make sure data gets proper repair in the next run and crash the app
